Split enums into separate module
